+++ src/dispatch/model.py
@@ -618,14 +618,21 @@ class DispatchModel(IOMixin):
                     LOGGER.warning("unable to write %s, %r", df_name, exc)
         return None, state
 
-    def __getattr__(self, item):
-        """Get values from ``self._metadata`` as if they were properties."""
-        try:
-            return self._metadata[item]
-        except KeyError:
-            raise AttributeError(
-                f"'{self.__class__.__qualname__}' object has no attribute '{item}'"
-            ) from None
+    @property
+    def name(self):  # noqa: D102
+        return self._metadata["name"]
+
+    @property
+    def version(self):  # noqa: D102
+        return self._metadata["version"]
+
+    @property
+    def created(self):  # noqa: D102
+        return self._metadata["created"]
+
+    @property
+    def jit(self):  # noqa: D102
+        return self._metadata["jit"]
 
     def set_metadata(self, key, value: Any) -> None:
         """Set a value in metadata."""